make cart link work
