+++ src/store/index.js
@@ -1,8 +1,8 @@
 import {createStore, applyMiddleware} from 'redux';
 import APImiddleware from '~src/middleware/API'
-import LocalStorageMiddleware from '~src/middleware/LocalStorage'
+import CartMiddleware from '~src/middleware/Cart'
 import reducers from '~src/reducers';
 
-const store = createStore(reducers, applyMiddleware(APImiddleware, LocalStorageMiddleware));
+const store = createStore(reducers, applyMiddleware(APImiddleware, CartMiddleware));
 
 export default store;
